+++ front-end/constants.js
@@ -1,0 +1,2 @@
+const BOOKS_PER_PAGE = 24;
+export { BOOKS_PER_PAGE };
