+++ front-end/index.html
@@ -16,7 +16,8 @@
       src="https://kit.fontawesome.com/a40723187e.js"
       crossorigin="anonymous"
     ></script>
-    <script type="module" src="./index.js" defer></script>
+    <script type="module" src="./init.js" defer></script>
+    <script type="module" src="./events.js" defer></script>
   </head>
   <body>
     <header>
